int to float in mutation rate

+++ NEUROMOD.py
@@ -9,7 +9,7 @@ class NEUROMOD():
     def __init__(self, n=50, theta=1, population_size=20, parents_size=10, children_size=10, max_generations=500, data='MNIST', phys='l1', individual_mutation=0.5, layerwise_mutation=0.5):
         self.theta, self.n = float(theta), int(n)
         self.population_size, self.parents_size, self.children_size, self.max_generations = int(population_size), int(parents_size), int(children_size), int(max_generations)
-        self.individual_mutation, self.layerwise_mutation = int(individual_mutation), int(layerwise_mutation)
+        self.individual_mutation, self.layerwise_mutation = float(individual_mutation), float(layerwise_mutation)
 
         self.objectives = {'acc': DatasetObjective(dataset_name = 'MNIST', evaluation_metric = 'acc'),
                            'phys': {'count': NetworkCostObjective(cost_metric = 'count'),
